fix(666): PoC api
