+++ GUI/src/components/chat/chat.module.scss
@@ -248,5 +248,6 @@
     position: fixed;
     bottom: 0;
     right: 0;
+    z-index: 20;
   }
 }
