+++ GUI/src/i18n/en/common.json
@@ -473,7 +473,7 @@
     "fileSignYesNo": "Do you want to sign this file? Yes/No",
     "redirectToCustomerSupport": "Conversation will be redirected to Customer Support",
     "waiting-for-user-input": "Waiting for user input",
-    "title": "Chat Integration",
+    "title": "TEST",
     "input-placeholder": "Enter your message..."
   },
   "objectEditor": {
