+++ GUI/src/i18n/et/common.json
@@ -474,7 +474,7 @@
     "fileSignYesNo": "Kas soovid faili allkirjastada? Jah/Ei",
     "redirectToCustomerSupport": "Vestlus suunatakse nõustajale",
     "waiting-for-user-input": "Ootan kasutaja sisendit",
-    "title": "Chat Integration",
+    "title": "TEST",
     "input-placeholder": "Kirjutage oma sõnum..."
   },
   "objectEditor": {
