+++ GUI/src/services/api.ts
@@ -1,6 +1,6 @@
-import axios, { AxiosError } from 'axios';
+import axios, { AxiosError, AxiosRequestConfig } from 'axios';
 
-const instance = axios.create({
+const defaultConfig = {
   baseURL: import.meta.env.REACT_APP_BASE_API_PATH,
   headers: {
     Accept: 'application/json',
@@ -8,7 +8,21 @@ const instance = axios.create({
     'Cache-Control': 'no-cache, no-store, must-revalidate',
   },
   withCredentials: true,
-});
+};
+
+const instance = axios.create(defaultConfig);
+
+export const createApiInstance = (additionalHeaders?: Record<string, string>) => {
+  const config: AxiosRequestConfig = {
+    ...defaultConfig,
+    headers: {
+      ...defaultConfig.headers,
+      ...additionalHeaders,
+    },
+  };
+
+  return axios.create(config);
+};
 
 instance.interceptors.response.use(
   (response) => {
